Struct xmodem::Xmodem

pub struct Xmodem<R> { /* fields omitted */ }

Implementation of the XMODEM protocol.

Methods

impl Xmodem<()>

pub fn transmit<R, W>(data: R, to: W) -> Result<usize> where
W: Read + Write,
R: Read,

Transmits data to the receiver to using the XMODEM protocol. If the length of the total data yielded by data is not a multiple of 128 bytes, the data is padded with zeroes and sent to the receiver.

Returns the number of bytes written to to, excluding padding zeroes.

pub fn transmit_with_progress<R, W>(
data: R,
to: W,
f: ProgressFn
) -> Result<usize> where
W: Read + Write,
R: Read,

Transmits data to the receiver to using the XMODEM protocol. If the length of the total data yielded by data is not a multiple of 128 bytes, the data is padded with zeroes and sent to the receiver.

The function f is used as a callback to indicate progress throughout the transmission. See the [Progress] enum for more information.

Returns the number of bytes written to to, excluding padding zeroes.

pub fn receive<R, W>(from: R, into: W) -> Result<usize> where
R: Read + Write,
W: Write,

Receives data from from using the XMODEM protocol and writes it into into. Returns the number of bytes read from from, a multiple of 128.

pub fn receive_with_progress<R, W>(
from: R,
into: W,
f: ProgressFn
) -> Result<usize> where
R: Read + Write,
W: Write,

Receives data from from using the XMODEM protocol and writes it into into. Returns the number of bytes read from from, a multiple of 128.

The function f is used as a callback to indicate progress throughout the reception. See the [Progress] enum for more information.

impl<T: Read + Write> Xmodem<T>

pub fn new(inner: T) -> Self

Returns a new Xmodem instance with the internal reader/writer set to inner. The returned instance can be used for both receiving (downloading) and sending (uploading).

pub fn new_with_progress(inner: T, f: ProgressFn) -> Self

Returns a new Xmodem instance with the internal reader/writer set to inner. The returned instance can be used for both receiving (downloading) and sending (uploading). The function f is used as a callback to indicate progress throughout the transfer. See the [Progress] enum for more information.

pub fn read_packet(&mut self, buf: &mut [u8]) -> Result<usize>

Reads (downloads) a single packet from the inner stream using the XMODEM protocol. On success, returns the number of bytes read (always 128).

The progress callback is called with Progress::Start when reception for the first packet has started and subsequently with Progress::Packet when a packet is received successfully.

Errors

Returns an error if reading or writing to the inner stream fails at any point. Also returns an error if the XMODEM protocol indicates an error. In particular, an InvalidData error is returned when:

The sender's first byte for a packet isn't EOT or SOH.

The sender doesn't send a second EOT after the first.

The received packet numbers don't match the expected values.

An error of kind Interrupted is returned if a packet checksum fails.

An error of kind ConnectionAborted is returned if a CAN byte is received when not expected.

An error of kind UnexpectedEof is returned if buf.len() < 128.

pub fn write_packet(&mut self, buf: &[u8]) -> Result<usize>

Sends (uploads) a single packet to the inner stream using the XMODEM protocol. If buf is empty, end of transmissions is sent. Users of this interface should ensure that write_packet(&[]) is called when data transmission is complete. On success, returns the number of bytes written.

The progress callback is called with Progress::Waiting before waiting for the receiver's NAK, Progress::Start when transmission of the first packet has started and subsequently with Progress::Packet when a packet is sent successfully.

Errors

Returns an error if reading or writing to the inner stream fails at any point. Also returns an error if the XMODEM protocol indicates an error. In particular, an InvalidData error is returned when:

The receiver's first byte isn't a NAK.

The receiver doesn't respond with a NAK to the first EOT.

The receiver doesn't respond with an ACK to the second EOT.

The receiver responds to a complete packet with something besides ACK or NAK.

An error of kind UnexpectedEof is returned if buf.len() < 128 && buf.len() != 0.

An error of kind ConnectionAborted is returned if a CAN byte is received when not expected.

An error of kind Interrupted is returned if a packet checksum fails.

pub fn flush(&mut self) -> Result<()>

Flush this output stream, ensuring that all intermediately buffered contents reach their destination.

Errors

It is considered an error if not all bytes could be written due to I/O errors or EOF being reached.